+++ src/main/java/de/alpharogroup/lang/ClassExtensions.java
@@ -348,8 +348,8 @@ public final class ClassExtensions
 	 * @throws URISyntaxException
 	 *             is thrown if a string could not be parsed as a URI reference.
 	 */
-	public static List<File> getDirectoriesFromResources(@NonNull String path, final boolean isPackage)
-		throws IOException, URISyntaxException
+	public static List<File> getDirectoriesFromResources(@NonNull String path,
+		final boolean isPackage) throws IOException, URISyntaxException
 	{
 		if (isPackage)
 		{
@@ -665,7 +665,8 @@ public final class ClassExtensions
 			if (url.getProtocol().equals("jar"))
 			{
 				InputStream resourceAsStream = ClassExtensions.getResourceAsStream(name, obj);
-				file = new File(System.getProperty("java.io.tmpdir"), obj.getClass().getSimpleName());
+				file = new File(System.getProperty("java.io.tmpdir"),
+					obj.getClass().getSimpleName());
 				Files.copy(resourceAsStream, file.toPath(), StandardCopyOption.REPLACE_EXISTING);
 			}
 			if (url.getProtocol().equals("file"))
@@ -687,7 +688,8 @@ public final class ClassExtensions
 	 *            The uri as String.
 	 * @return The InputStream from the uri.
 	 */
-	public static InputStream getResourceAsStream(final @NonNull Class<?> clazz, final @NonNull String uri)
+	public static InputStream getResourceAsStream(final @NonNull Class<?> clazz,
+		final @NonNull String uri)
 	{
 		InputStream is = clazz.getResourceAsStream(uri);
 		if (null == is)
@@ -720,7 +722,8 @@ public final class ClassExtensions
 	 *            The Object.
 	 * @return The resource or null if the resource does not exists.
 	 */
-	public static InputStream getResourceAsStream(final @NonNull String name, final @NonNull Object obj)
+	public static InputStream getResourceAsStream(final @NonNull String name,
+		final @NonNull Object obj)
 	{
 		InputStream inputStream = obj.getClass().getResourceAsStream(name);
 		if (null == inputStream)