+++ src/test/java/de/alpharogroup/reflection/ReflectionExtensionsTest.java
@@ -49,15 +49,15 @@ public class ReflectionExtensionsTest
 {
 
 	/**
-	 * Test method for {@link ReflectionExtensions#copyArray(Object)}
+	 * Test method for {@link ReflectionExtensions#copyArray(Object[])}
 	 */
 	@Test
 	public void testCopyArray()
 	{
 		Integer[] actual;
 		Integer[] expected;
-		actual = ArrayFactory.newArray(1, 2, 3);
-		expected = (Integer[])ReflectionExtensions.copyArray(actual);
+		expected = ArrayFactory.newArray(1, 2, 3);
+		actual = ReflectionExtensions.copyArray(expected);
 		assertTrue(Arrays.deepEquals(actual, expected));
 	}